// app/src/models/address.ts
export interface Province {
    provincekey: number;
    provincenameth: string;
    provinceshort: string;
}

export interface District {
    districtkey: number;
    districtnameth: string;
}

export interface SubDistrict {
    subdistrictkey: number;
    subdistrictnameth: string;
}

export interface ZipCode {
    zipcode: string;
}